feat: contract dropdown custom, asset ID dropdown custom
