+++ src/components/TransactionForms/FormInput/index.tsx
@@ -3,10 +3,12 @@ import { setQueryAndRouter } from '@/utils';
 import { NextParsedUrlQuery } from 'next/dist/server/request-meta';
 import dynamic from 'next/dynamic';
 import { NextRouter, useRouter } from 'next/router';
-import { ChangeEventHandler, useEffect, useRef } from 'react';
+import { ChangeEventHandler, useEffect, useRef, useState } from 'react';
 import { FieldValues, useFormContext, UseFormGetValues } from 'react-hook-form';
 import {
   Container,
+  DropdownCustomLabel,
+  DropdownCustomLabelSelectStyles,
   ErrorMessage,
   InfoIcon,
   InputLabel,
@@ -21,6 +23,11 @@ import {
 } from './styles';
 
 const Select = dynamic(() => import('./Select'), {
+  ssr: false,
+  loading: () => null,
+});
+
+const DropdownCustomLabelSelect = dynamic(() => import('@/components/Select'), {
   ssr: false,
   loading: () => null,
 });
@@ -55,6 +62,12 @@ export interface IFormInputProps extends IBaseFormInputProps {
 export interface ICustomFormInputProps extends IBaseFormInputProps {
   onChange: ChangeEventHandler<any>;
 }
+
+export const customDropdownOptions = [
+  { label: 'No', value: 'no' },
+  { label: 'Text', value: 'text' },
+  { label: 'Number', value: 'number' },
+];
 
 export const onChangeWrapper = (
   isMultiContract: boolean,
@@ -117,6 +130,7 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
   ...rest
 }) => {
   const areaRef = useRef<HTMLTextAreaElement | null>(null);
+  const [isCustom, setIsCustom] = useState(customDropdownOptions[0]);
   const router = useRouter();
   const { isMultiContract } = useMulticontract();
   const {
@@ -144,7 +158,7 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
 
   const registerProps =
     name &&
-    (type !== 'number'
+    (type !== 'number' && isCustom.value !== 'number'
       ? register(name, {
           required: {
             value: required || false,
@@ -172,6 +186,10 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
             customOnChange,
           ),
           validate: (value: any) => {
+            if (Number.isNaN(value)) {
+              return 'Only numbers allowed';
+            }
+
             if (value < (min || 0)) {
               return `Minimum value is ${min || 0}`;
             }
@@ -208,7 +226,7 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
     return () => {
       name && unregister(name);
     };
-  }, [name, unregister]);
+  }, [name, unregister, isCustom]);
 
   const areaProps = {
     error: Boolean(error),
@@ -247,7 +265,7 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
     e.target.blur();
   };
 
-  type === 'number' &&
+  (type === 'number' || isCustom.value === 'number') &&
     (inputProps = {
       ...inputProps,
       step: 1 / 10 ** precision,
@@ -282,7 +300,7 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
         <InputLabel disabled={inputProps.disabled}>
           <span>
             {title || name}
-            {required && <RequiredSpan> (required)</RequiredSpan>}
+            {required && <RequiredSpan>(required)</RequiredSpan>}
           </span>
           {tooltip && (
             <TooltipContainer>
@@ -291,6 +309,20 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
                 <span>{tooltip}</span>
               </TooltipContent>
             </TooltipContainer>
+          )}
+          {type === 'dropdown' && (
+            <DropdownCustomLabel>
+              <span>Custom value?</span>
+              <DropdownCustomLabelSelect
+                options={customDropdownOptions}
+                value={isCustom}
+                onChange={(e: any) => {
+                  setIsCustom(e) as any;
+                }}
+                styles={DropdownCustomLabelSelectStyles}
+                isSearchable={false}
+              />
+            </DropdownCustomLabel>
           )}
         </InputLabel>
       )}
@@ -306,7 +338,9 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
           </ToggleContainer>
         </>
       )}
-      {type === 'dropdown' && <Select {...selectProps} />}
+      {type === 'dropdown' && isCustom.value === 'no' && (
+        <Select {...selectProps} />
+      )}
       {type === 'textarea' && (
         <StyledTextArea
           onKeyDown={handleKey}
@@ -320,7 +354,7 @@ const FormInput: React.FC<IFormInputProps | ICustomFormInputProps> = ({
       )}
       {type === 'hidden' && <StyledInput {...inputProps} />}
       {type !== 'checkbox' &&
-        type !== 'dropdown' &&
+        (type !== 'dropdown' || isCustom.value !== 'no') &&
         type !== 'textarea' &&
         type !== 'hidden' && <StyledInput {...inputProps} />}
