+++ src/components/TransactionForms/FormInput/styles.ts
@@ -308,6 +308,8 @@ export const InputLabel = styled.label<ILabel>`
 
   transition: transform 0.2s ease;
 
+  z-index: 3;
+
   > span,
   p {
     opacity: ${props => (props.disabled ? 0.6 : 1)};
@@ -340,3 +342,44 @@ export const ErrorMessage = styled.span<{ warning?: boolean }>`
       color: ${theme.status.warning} !important;
     `}
 `;
+
+export const DropdownCustomLabel = styled(InputLabel)`
+  margin: 0;
+  > span {
+    margin-left: auto;
+  }
+`;
+
+export const DropdownCustomLabelSelectStyles = {
+  control: (base: any): any => ({
+    ...base,
+    height: '22px !important',
+    minWidth: '4rem',
+    minHeight: '22px !important',
+    padding: '0 !important',
+    fontSize: '0.8rem',
+    lineHeight: '0.8rem',
+  }),
+  valueContainer: (base: any): any => ({
+    ...base,
+    padding: '0 0.25rem 0 0.5rem',
+    margin: '-0.35rem 0 0 0 ',
+    lineHeight: '0.8rem',
+    height: '2rem',
+  }),
+  indicatorSeparator: (base: any): any => ({
+    ...base,
+    display: 'none',
+  }),
+  dropdownIndicator: (base: any): any => ({
+    ...base,
+    padding: '0 0.5rem 0 0',
+    margin: '-0.35rem 0 0 0 ',
+  }),
+  option: (base: any): any => ({
+    ...base,
+    fontSize: '0.8rem',
+    lineHeight: '0.8rem',
+    padding: '0.25rem',
+  }),
+};
